Initial commit for the Deep Griffin Lim Model
The Griffin-Lim Algorithm is a method for reconstructing a signal from
its magnitude spectrogram, while making an educated guess about the
phase. It starts with a random phase, then iteratively refines the
phase estimate by applying a constraint in the time domain and then
projecting back to the frequency domain. The process is repeated until
the algorithm converges to a stable solution, providing a reconstructed
signal that matches the given magnitude spectrogramas.

Deep Griffin Lim is a denoising model that takes the fundamental inter
ative apporoach from the GL algorithm but uses a CNN to calculate a res
idual map that is then subtracted from a noisy signal to produce a
reconstructed phase signal