Add testing functionality and health dashboard
This update adds testing functionality and a health dashboard to check
the dimensionality of the data that goes into the model and out of the
model. This was done to check the system easily and reduce debugging
time.

The testing functionality comprises of two systems. Every epoch the res
idual, the clear image the final image and the consistency image (z~)
are being tracked by tensorboard.

In addition, the last batch of the testing set is converted back to
audible audio and saved in the out folder.