Add functionality clipped samples correctly
The new functionality sets a global clipping point in the signal the
first time a file is clipped. This global offset is then used for all
files. This was done to ensure that the files in the data loaders are
always the same and are not being changed every time the getitem
function is called.